Connection string moved to SpatiaLite.
I would prefer to have made it static, but the super-
class doesn't let me do that, and it's useful to have
access to it in the writer. We could also create a
driver in the writer to retrieve it, but instantiating a
class only for retrieving a string is super-lame.